metrics/requirements.txt
# Philips Internal GitHub Metrics Collection
# Specialized dependencies for synergy-base repository and hmp-synergy-blr team

# Core Airflow for automation
apache-airflow>=2.10.0
apache-airflow-providers-http>=4.7.0

# GitHub API integration for Philips Internal
PyGithub==2.1.1
requests==2.31.0

# Data processing for metrics calculation
pandas==2.1.4
numpy==1.24.3
python-dateutil==2.8.2

# Visualization for Philips dashboards  
plotly==5.17.0
dash>=2.15.0
dash-bootstrap-components>=1.5.0
streamlit>=1.28.0
dash==2.16.1
dash-bootstrap-components==1.5.0

# Configuration and utilities
python-dotenv==1.0.0
pydantic==2.5.2
click==8.1.7

# Testing
pytest==7.4.3
pytest-mock==3.12.0

# Development
black==23.12.1
flake8==6.1.0
pre-commit==3.6.0
